requirements.txt
# Core dependencies (REQUIRED)
Pillow==10.1.0                  # Image manipulation
opencv-python==4.8.1.78         # Computer vision for smart cropping
numpy==1.24.3                   # Numerical operations
requests==2.31.0                # HTTP requests for API calls
pyyaml==6.0.1                   # YAML file parsing
openai==1.3.0                   # OpenAI API for DALL-E 3 image generation (PRIMARY)

# API Framework (REQUIRED for API server)
fastapi==0.104.1                # REST API framework
uvicorn[standard]==0.24.0       # ASGI server
pydantic==2.5.0                 # Data validation

# Image Processing Enhancement (OPTIONAL but recommended)
opencv-contrib-python==4.8.1.78 # Extended OpenCV features

# Testing (OPTIONAL - for development)
pytest==7.4.3                   # Testing framework
pytest-cov==4.1.0               # Coverage reporting
pytest-asyncio==0.21.1          # Async test support

# Development Tools (OPTIONAL)
black==23.11.0                  # Code formatting
mypy==1.7.0                     # Type checking
pylint==3.0.2                   # Code linting

# Demo UI (OPTIONAL - for interactive demo)
streamlit==1.28.2               # Web UI framework
pandas==2.0.3                   # Data manipulation for UI
plotly==5.17.0                  # Interactive charts

# Production Features (OPTIONAL but recommended)
structlog==23.2.0               # Structured logging
aiofiles==23.2.1                # Async file operations
httpx==0.25.1                   # Async HTTP client
backoff==2.2.1                  # Retry logic (OPTIONAL)
redis==5.0.1                    # Queue/caching (OPTIONAL)
celery==5.3.4                   # Task queue (OPTIONAL)

# Additional GenAI Provider SDKs (OPTIONAL - for multi-provider support)
stability-sdk==0.8.1            # Stability AI API (alternative provider)
replicate==0.15.0               # Replicate API (alternative provider)
anthropic==0.7.0                # Anthropic Claude API (for future enhancements)

# Cloud Storage (OPTIONAL - for production)
boto3==1.28.0                   # AWS S3 storage
azure-storage-blob==12.19.0    # Azure Blob storage
google-cloud-storage==2.10.0   # Google Cloud Storage

# Monitoring & Observability (OPTIONAL - for production)
prometheus-client==0.19.0       # Metrics export
opentelemetry-api==1.21.0      # Distributed tracing
sentry-sdk==1.38.0              # Error tracking

# Database (OPTIONAL - for production)
psycopg2-binary==2.9.9          # PostgreSQL
sqlalchemy==2.0.23              # ORM
alembic==1.12.1                 # Database migrations
